pyproject.toml
[project]
name = "lomnia-ingester"
version = "0.0.1"
description = "Coordinate, ingest and save data from multiple extractors"
authors = [{ name = "Lorenzo Piccoli", email = "lorenzopicoli@me.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.9,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "boto3>=1.42.5",
    "boto3-stubs[s3]>=1.42.5",
    "pika>=1.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[project.urls]
Homepage = "https://lorenzopicoli.github.io/lomnia-ingester/"
Repository = "https://github.com/lorenzopicoli/lomnia-ingester"
Documentation = "https://lorenzopicoli.github.io/lomnia-ingester/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",

    "ty>=0.0.1a16",

    "ruff>=0.11.5",

]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lomnia_ingester"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.9"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
